fix: interactive mode not working for csv and list inputs
